+++ makecode.py
@@ -1,4 +1,6 @@
-#!/usr/bin/env python3 
+#!/usr/bin/env python3
+
+"""Auto-generate terrascript/PROVIDER/{r,d}.py files."""
 
 import glob
 import collections
@@ -6,42 +8,56 @@ import os
 import os.path
 import sys
 import time
+import re
+import github3
+import base64
 
 PREFIX = 'terrascript'
+REPO_OWNER = 'terraform-providers'
+PROVIDERS = [
+    'aws',
+    'azurerm',
+    'google',
+    'template',
+    'openstack',
+    'kubernetes',
+    'docker'
+]
+PROVIDERS.sort()
+
+
+REGEX = re.compile(b'".*?_(?P<type_>.+)":\s+(?P<class_>resource|data)')
+#"azurerm_traffic_manager_profile":   resourceArmTrafficManagerProfile(),
+
+try:
+    # Travis-CI
+    TOKEN = os.environ['GITHUB_TOKEN']
+except KeyError:
+    # @home
+    TOKEN = open('GITHUB_TOKEN').read().strip()
 
 thisdir = os.path.abspath('.')
 if not os.path.isdir(os.path.join(thisdir, PREFIX)):
     print('Script must be run from python-terrascript folder')
     sys.exit(1)
 
-paths = [p for p in glob.glob('terraform-providers/*/*/*.go')
-        if not p.endswith('test.go') and ('/resource_' in p or '/data_source_' in p)]
 
-paths.sort()
+gh = github3.login(token=TOKEN)
+for provider in PROVIDERS:
+    print(provider)
+    reponame = 'terraform-provider-{}'.format(provider)
+    repo = gh.repository(REPO_OWNER, reponame)
+    branch = repo.branch(repo.default_branch)
+    commit = branch.commit
+    tree = repo.tree(commit.sha)
 
-# Dictionary as {PROVIDER: ([RESOURCE, ...],[DATA, ...])}
-data = collections.defaultdict(lambda : ([],[]))
+    regex = re.compile('^'+provider+'/provider.go')
+    provider_go_hash = [h for h in tree.recurse().tree
+                       if regex.match(h.path)][0]
 
-for path in paths:
-    x,y,provider,klass = path.split('/')
+    provider_go_blob = repo.blob(provider_go_hash.sha)
+    decoded = base64.b64decode(provider_go_blob.content)
 
-    # Strip '.go' suffix
-    klass = klass.split('.')[0]
-
-    if klass.startswith('resource_'):
-        klass = klass[len('resource_'):]
-        if not klass.startswith(provider + '_'):
-            klass = provider + '_' + klass
-        data[provider][0].append(klass)
-    else:
-        klass = klass[len('data_source_'):]
-        if not klass.startswith(provider + '_'):
-            klass = provider + '_' + klass
-        data[provider][1].append(klass)
-
-# Create Python modules
-#
-for provider,v in data.items():
     provider_path = os.path.join(thisdir, PREFIX, provider)
     provider_init_path = os.path.join(thisdir, PREFIX, provider, '__init__.py')
     resource_path = os.path.join(provider_path, 'r.py')
@@ -49,21 +65,24 @@ for provider,v in data.items():
 
     if not os.path.exists(provider_path):
         os.mkdir(provider_path)
-        # TODO: create __init__.py
 
     with open(provider_init_path, 'w') as fp:
         fp.write('"""' + time.strftime('%Y-%m-%d %H:%M:%S') + '"""' + '\n')
 
-    print(resource_path)
-    with open(resource_path, 'w') as fp:
-        fp.write('from terrascript import _resource\n')
-        for resource in v[0]:
-            fp.write('class {}(_resource): pass\n'.format(resource))
-            fp.write('{} = {}\n\n'.format(resource[len(provider)+1:], resource))
+    with open(resource_path, 'w') as fpr:
+        fpr.write('from terrascript import _resource\n')
 
-    print(data_path)
-    with open(data_path, 'w') as fp:
-        fp.write('from terrascript import _data\n')
-        for data in v[1]:
-            fp.write('class {}(_data): pass\n'.format(data))
-            fp.write('{} = {}\n\n'.format(data[len(provider)+1:], data))
+        with open(data_path, 'w') as fpd:
+            fpd.write('from terrascript import _data\n')
+
+            for m in REGEX.finditer(decoded):
+                type_ = m.group('type_').decode('utf-8')
+                class_ = m.group('class_').decode('utf-8')
+
+                if class_ == 'resource':
+                    fpr.write('class {}_{}(_resource): pass\n'.format(provider, type_))
+                    fpr.write('{} = {}_{}\n\n'.format(type_, provider, type_))
+                else:
+                    fpd.write('class {}_{}(_data): pass\n'.format(provider, type_))
+                    fpd.write('{} = {}_{}\n\n'.format(type_, provider, type_))
+