+++ terrascript/terraform/d.py
@@ -1,0 +1,3 @@
+from terrascript import _data
+class terraform_remote_state(_data): pass
+remote_state = terraform_remote_state