+++ tests/test.py
@@ -164,7 +164,8 @@ class TestProvider(_Validate):
         provider('NAME', region='REGION')
         
         
-class TestTerraform(_Validate):
+class Test0Terraform(_Validate):
+    # These tests must be executed first, thus the '0' in the name
     def test(self):
         terraform(required_version='> 0.9')
         
@@ -197,4 +198,12 @@ class TestFunctions(_Validate):
         aws_region = variable('aws_region', default='us-east-1', description='The AWS region to create things in.')
         aws_amis = variable('aws_amis', default={'us-east-1': 'ami-5f709f34', 'us-west-2': 'ami-7f675e4f'})
         assert function.lookup(aws_amis, aws_region) == '${lookup(var.aws_amis,var.aws_region)}'
+        
+        
+class TestResourceData(_Validate):
+    def test_data(self):
+        data('aws_instance', 'NAME')
+        
+    def test_resource(self):
+        resource('aws_instance', 'NAME')