+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.2.0] - 2017-07-26
+### Added
+* `google` provider.
+* `azurerm` provider.
+* `terraform` provider.
+* `template` provider.
+* `openstack` provider.
+* `kubernetes` provider.
+* `docker` provider.
+
 ## [0.1.0] - 2017-07-25
 * First release.
 * Supports only AWS provider.